Refactor `FallbackImage`
