+++ src/shared/components/AvatarImage/index.tsx
@@ -1,10 +1,11 @@
 import React, { useMemo } from 'react'
 import styled from 'styled-components'
 import { isAddress } from 'web3-validator'
-import UnstyledFallbackImage from '~/shared/components/FallbackImage'
+import { FallbackImage as UnstyledFallbackImage } from '~/components/FallbackImage'
 import SvgIcon from '~/shared/components/SvgIcon'
 import Initials from './Initials'
 import Identicon from './Identicon'
+
 const FallbackImage = styled(UnstyledFallbackImage)`
     display: block;
     height: 100%;
