--- src/shared/components/FallbackImage/fallbackImage.pcss
@@ -1,21 +1,0 @@
-.defaultImagePlaceholder {
-  position: relative;
-  background-image: linear-gradient(135deg, #0045FF 0%, #7200EE 100%);
-  border-radius: 2px;
-  transition: 50ms ease-in-out transform;
-
-  img {
-    width: 100%;
-    height: 100%;
-    background-color: transparent;
-  }
-
-  svg {
-    position: absolute;
-    width: 80px;
-    height: 80px;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-  }
-}
